refactor twitter embed
